+++ nemo/collections/asr/losses/__init__.py
@@ -15,7 +15,7 @@
 from nemo.collections.asr.losses.angularloss import AngularSoftmaxLoss
 from nemo.collections.asr.losses.ctc import CTCLoss
 from nemo.collections.asr.losses.lattice_losses import LatticeLoss
-from nemo.collections.asr.losses.pt_losses.contrastive import ContrastiveLoss
-from nemo.collections.asr.losses.pt_losses.ctc import CTCLossForSSL
-from nemo.collections.asr.losses.pt_losses.mlm import MLMLoss
-from nemo.collections.asr.losses.pt_losses.rnnt import RNNTLossForSSL
+from nemo.collections.asr.losses.ssl_losses.contrastive import ContrastiveLoss
+from nemo.collections.asr.losses.ssl_losses.ctc import CTCLossForSSL
+from nemo.collections.asr.losses.ssl_losses.mlm import MLMLoss
+from nemo.collections.asr.losses.ssl_losses.rnnt import RNNTLossForSSL
